+++ src/index.js
@@ -2,14 +2,14 @@ import './init';
 
 export {VERSION} from './utils/globals';
 
-export {default as Log} from './log';
+export {default as Log} from './log/log';
 
 export {
   default as Bench,
   logResultsAsMarkdownTable,
   logResultsAsTree,
   logResultsAsTreeWithElapsed
-} from './bench';
+} from './bench/bench';
 
 // // Make Probe available as global variable for debugging purposes
 // // TODO - convenient but can be a concern for "purists"
@@ -17,6 +17,8 @@ export {
 // import Probe from './probe';
 // import {global} from './utils/globals';
 // global.Probe = Probe;
+
+export {default as NodeTestDriver} from './test/node-test-driver';
 
 // Experimental exports
 import {enableDOMLogging} from './utils/log-to-dom';